pyproject.toml
[tool.poetry]
name = "calendar-back"
version = "0.1.0"
description = ""
authors = ["byungyong <quddyddlrk@naver.com>"]
readme = "README.md"
packages = [{include = "calendar_back"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
